Updated pouchdb-auth so that we can publish a scoped version for use with express-pouchdb.
