+++ package.json
@@ -11,7 +11,7 @@
     "release": "./bin/release.sh"
   },
   "dependencies": {
-    "base64url": "^2.0.0",
+    "base64url": "^3.0.0",
     "basic-auth": "^2.0.0",
     "bluebird": "^3.4.7",
     "body-parser": "^1.16.1",
@@ -20,7 +20,7 @@
     "compression": "^1.6.2",
     "cookie-parser": "^1.4.3",
     "corser": "~2.0.0",
-    "couchdb-calculate-session-id": "^1.1.0",
+    "couchdb-calculate-session-id": "^1.1.3",
     "couchdb-log-parse": "^0.0.4",
     "crypto-lite": "^0.2.0",
     "denodeify": "^1.2.1",
@@ -36,7 +36,7 @@
     "nomnom": "^1.8.1",
     "object-assign": "^4.1.0",
     "on-finished": "^2.3.0",
-    "@the-t-in-rtf/pouchdb-all-dbs": "1.0.2",
+    "@the-t-in-rtf/pouchdb-all-dbs": "1.0.2-dev.20181211T110844Z.64f2e71",
     "pouchdb-fauxton": "^0.0.6",
     "promise-nodify": "^1.0.2",
     "random-uuid-v4": "0.0.6",
